Move settings to a common library (trying to figure out how to start modularizing things)
