+++ SubLinkCommon/SubLinkCommon.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <Configurations>Debug;Release</Configurations>
+        <Platforms>AnyCPU</Platforms>
+        <RootNamespace>xyz.yewnyx.SubLink</RootNamespace>
+        <Authors>yewnyx</Authors>
+        <Product>SubLinkCommon</Product>
+        <Description>SubLink common libraries</Description>
+        <Copyright>©2022-present Yewnyx</Copyright>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
+    </ItemGroup>
+</Project>
