+++ SubLinkKick/SubLinkKick.csproj
@@ -7,7 +7,7 @@
         <Configurations>Debug;Release</Configurations>
         <Platforms>AnyCPU</Platforms>
         <ApplicationIcon>..\SubLink.ico</ApplicationIcon>
-        <RootNamespace>xyz.yewnyx</RootNamespace>
+        <RootNamespace>xyz.yewnyx.SubLink</RootNamespace>
         <Authors>yewnyx</Authors>
         <Product>SubLinkKick</Product>
         <Description>SubLink programmable streamer app (Kick Version)</Description>
@@ -42,5 +42,9 @@
       <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
       <PackageReference Include="VRCOscLib" Version="1.4.3" />
       <PackageReference Include="XSNotifications" Version="0.1.2" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\SubLinkCommon\SubLinkCommon.csproj" />
     </ItemGroup>
 </Project>
